test: :white_check_mark: rename ReactNativeNavigationScenario
