+++ test/react-native/features/fixtures/scenario-launcher/scenarios/ReactNativeNavigationScenario.js
